fix: satisfy hao observation lint checks

+++ backend/internal/haocli/observe_unix.go
@@ -24,5 +24,5 @@ func diskAvailable(path string) (uint64, error) {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return 0, err
 	}
-	return uint64(stat.Bavail) * uint64(stat.Bsize), nil
+	return stat.Bavail * uint64(stat.Bsize), nil
 }
